UI Updates + River Generation
While trying to debug my river generation, I realized having a way to view the coords and altitude of my terrain would be nice, so I updated my UI.

I've realized some of my logic was flawed with the river generation, and it's closer than ever to being in a state I'm happy with. I still need to work out a way to keep a river from merging with itself.