Added temperature and moisture maps
I've added some temperature and moisture into the terrain generation, and I've also made it so that there is a temperature view, a altitude view, and a humidity view. I think some things like mountain generation will have to change, and it might be good to go into generating fault lines next.