Removed Temperature Noise
I feel like the blend of 3 different noise maps was too harsh, so I've decided to focus biomes on altitude and moisture. Altitude kind of doubles as temperature since the higher you go the colder it gets. It's looking a bit better, but I think it will look even better when I change the moisture map to be influenced by nearby water, rain shadow maybe, and other natural processes.