Updated Moisture Map Based on Tectonics
I made a sub-class for my tectonic plates and hopefully this will expand functionality later on. For now, I made it so that the moisture is largely continent based and thus biomes are continent based. I still have a lot to do, and  I'm not sure if this is how I'll do it later on.